adds eotsd version
